+++ visualization/app/codeCharta/state/store/appStatus/currentFilesAreSampleFiles/currentFilesAreSampleFiles.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from "@ngrx/store"
+
+export const setCurrentFilesAreSampleFiles = createAction("CURRENT_FILES_ARE_SAMPLE_FILES", props<{ value: boolean }>())